+++ src/anomalib/deploy/__init__.py
@@ -3,10 +3,7 @@
 # Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
+from .export import ExportType
 from .inferencers import Inferencer, OpenVINOInferencer, TorchInferencer
 
-__all__ = [
-    "Inferencer",
-    "OpenVINOInferencer",
-    "TorchInferencer",
-]
+__all__ = ["Inferencer", "OpenVINOInferencer", "TorchInferencer", "ExportType"]
