+++ src/anomalib/models/components/__init__.py
@@ -3,7 +3,7 @@
 # Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .base import AnomalyModule, BufferListMixin, DynamicBufferMixin, ExportType, MemoryBankMixin
+from .base import AnomalyModule, BufferListMixin, DynamicBufferMixin, MemoryBankMixin
 from .dimensionality_reduction import PCA, SparseRandomProjection
 from .feature_extractors import TimmFeatureExtractor, TorchFXFeatureExtractor
 from .filters import GaussianBlur2d
@@ -23,5 +23,4 @@ __all__ = [
     "SparseRandomProjection",
     "TimmFeatureExtractor",
     "TorchFXFeatureExtractor",
-    "ExportType",
 ]
