+++ tests/integration/model/test_models.py
@@ -13,8 +13,9 @@ import pytest
 
 from anomalib import TaskType
 from anomalib.data import AnomalibDataModule, MVTec
+from anomalib.deploy import ExportType
 from anomalib.engine import Engine
-from anomalib.models import AnomalyModule, ExportType, get_available_models, get_model
+from anomalib.models import AnomalyModule, get_available_models, get_model
 
 
 def models() -> set[str]:
