+++ tox.ini
@@ -28,7 +28,7 @@ deps =
     pytest-cov
     pytest-mock
     pytest-order
-    pytest-rerunfailures
+    flaky
     nbmake
 
 commands =
